requirements.txt
# Django and Core
Django==5.2.4
djangorestframework>=3.14.0

# Gemini AI
google-generativeai>=0.3.0

# Razorpay Integration
razorpay>=1.3.0

# QR Code Generation
qrcode>=7.4.2
Pillow>=9.5.0

# Email Backend
django-anymail>=10.0

# Environment Variables (optional but recommended)
python-dotenv>=1.0.0
